+++ engine/webbunjs/src/server.js
@@ -63,12 +63,7 @@ module.exports = (...args) => {
           setting;
         const { savestore, store, verbose, ...setsession } =
           setting[setting.args.engine].session;
-        const {
-          assets,
-          atomic,
-          webaction: reaction,
-          utilities,
-        } = compmgr.assist;
+        const { assets, atomic, reaction, utilities } = compmgr.assist;
 
         try {
           // Setup server log